+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/rest/service/RestPostDecoderAction.java
@@ -111,7 +111,7 @@ public class RestPostDecoderAction extends AbstractCreateAction {
         RestResponse engineValidation = this.engine.validateResource(Constants.KEY_DECODER, resource);
         if (engineValidation.getStatus() != RestStatus.OK.getStatus()) {
             return new RestResponse(
-                    "Engine Validation Failed: " + engineValidation.getMessage(),
+                    Constants.E_400_ENGINE_VALIDATION_FAILED + " " + engineValidation.getMessage(),
                     RestStatus.BAD_REQUEST.getStatus());
         }
         return null;